stop supporting Unix-like operating systems
