--- src/utils/exec-async.js
@@ -1,7 +1,0 @@
-"use strict";
-
-const { exec } = require("child_process");
-const util = require("util");
-const execAsync = util.promisify(exec);
-
-module.exports = execAsync;
