+++ src/utils/exec-file-async.js
@@ -1,17 +1,7 @@
 "use strict";
 
 const { execFile } = require("child_process");
-
-const execFileAsync = (file, args = [], callback) => {
-  return new Promise((resolve, reject) => {
-    execFile(file, args, (error, stdout) => {
-      if (error) {
-        reject(error);
-        return;
-      }
-      resolve(callback ? callback(stdout) : stdout);
-    });
-  });
-};
+const util = require("util");
+const execFileAsync = util.promisify(execFile);
 
 module.exports = execFileAsync;
