--- src/win32/cancel.js
@@ -1,7 +1,0 @@
-"use strict";
-
-const cancel = (jobId, options = {}) => {
-  throw new Error("The method is not yet implemented for win32");
-};
-
-module.exports = cancel;
