--- src/win32/observe.js
@@ -1,7 +1,0 @@
-"use strict";
-
-const observe = (jobId, timeout = 60000, delay = 1000, options = {}) => {
-  throw new Error("The method is not yet implemented for win32");
-};
-
-module.exports = observe;
