+++ webpack.production.config.js
@@ -13,7 +13,7 @@ module.exports = {
   plugins: [
     new CleanWebpackPlugin(),
     new CopyPlugin([
-      { from: "./src/win32/SumatraPDF.exe" },
+      { from: "./src/SumatraPDF.exe" },
       { from: "./src/types.d.ts" },
     ]),
   ],
